+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 4.0.1
+### General
+* Refactored `pdf_combiner` to use `compute` for native calls (`MethodChannel`), reducing potential UI freezes (UI thread blocking).
+
 ## 4.0.0
 ### General
 * Errors are being recovered from native and sent through the message.